test(redirect): fold the benchmarks into redirect_test.go
Its imports were a subset of the test file's, so the split bought nothing.
